Added advertisement pages
The ad system is something which is in heavy development and most likely
won't be complete when the main rewrite is done.
The terms/rules for ads laid out in these pages may change in the future
but can be used as a **rough** guideline. The prices may change also.
The current plan for ads is to have one slot at the top of the page and
one at the bottom, one ad per user with gifs allowed (manually approved)
to avoid annoying ads and make the experience pleasant for our visitors.
Also advertisers will receive statistics about their campaign and how it
is doing.

Advertisements are planned to help fund mc-api, as it is provided as a
free service without guarantee or warranty and is very expensive to run,
currently being funded by me.

If you have any comments or feedback please leave a comment! :smiley: