Relaxed api rules for cache
If something is cached then the server will not check the api key given
or require one at all.
This should make cached responses optimum speed, however if the request
is not in cache then an api key will be required (unless stated
otherwise in the documentation)

This may seem like a silly thing to do or have but there are niche use
cases for it.
For example: Someone is logged into your server, you queried our api
about them when they joined. Your website uses javascript which queries
our api too, instead of putting your api key in the public javascript
(which isn't a good idea) we can return the cached response without the
need for an api key.